+++ django_alt/managers.py
@@ -51,13 +51,16 @@ class ValidatedManager:
         self.validator.will_create_or_update()
         self.validator.base_db()
 
-        instance = self.model_class.objects.create(**self.validator.attrs)
+        instance = self.create_instance()
 
         self.validator.did_create(instance)
         self.validator.did_create_or_update(instance)
         self.validator.post()
 
         return instance
+
+    def create_instance(self):
+        return self.model_class.objects.create(**self.validator.attrs) if self.model_class is not None else None
 
     def do_update(self, instance, **attrs):
         """
@@ -78,14 +81,18 @@ class ValidatedManager:
         self.validator.will_create_or_update()
         self.validator.base_db()
 
-        for k, v in self.validator.attrs.items():
-            setattr(instance, k, v)
-        instance.save()
+        instance = self.update_instance(instance)
 
         self.validator.did_update(instance)
         self.validator.did_create_or_update(instance)
         self.validator.post()
 
+        return instance
+
+    def update_instance(self, instance):
+        for k, v in self.validator.attrs.items():
+            setattr(instance, k, v)
+        instance.save()
         return instance
 
     def _do_delete_pre(self, queryset, **attrs):
@@ -100,6 +107,14 @@ class ValidatedManager:
         self.validator.post()
         return queryset
 
+    def delete_instance(self, instance):
+        assert instance is not None, (
+            'An instance was not passed to `{}` for deletion\n'
+            'but `delete` was called.'
+            'Did you forget to include a `query` clause in an endpoint `delete` config?'
+        ).format(self.__class__.__qualname__)
+        return instance.delete()
+
     def do_delete(self, queryset, **attrs):
         """
         Validates given attribute dict and calls delete on given queryset.
@@ -108,6 +123,6 @@ class ValidatedManager:
         :return: deleted instance(s)
         """
         self._do_delete_pre(queryset, **attrs)
-        queryset.delete()
+        queryset = self.delete_instance(queryset)
         self._do_delete_post(queryset, **attrs)
         return queryset